added quadrotor
